requirements.txt
streamlit
python-pptx
Pillow
google-genai